mitosheet: cleanup minor bug fixes per review
