+++ mitosheet/src/components/elements/Dropdown.tsx
@@ -7,7 +7,6 @@ import '../../../css/elements/Dropdown.css'
 import { useCallOnAnyClick } from '../../hooks/useCallOnAnyClick';
 import { classNames } from '../../utils/classNames';
 import { fuzzyMatch } from '../../utils/strings';
-import DropdownCheckmark from '../icons/DropdownCheckmark';
 import Row from '../spacing/Row';
 import Input from './Input';
 import { WIDTH_MAPPING } from './sizes.d';
@@ -84,6 +83,10 @@ type BoundingRect = {
 /* 
     Helper function ensures that an item is visible inside a container, 
     and scrolls the minimal amount to make sure that it is visible.
+
+    NOTE: container must have some non-default positioning on it for 
+    this function to work properly, otherwise scrolling will look all
+    screwed up and you'll be confused for a while.
 
     The topAdjustment is the amount of space you can leave at the top of 
     the container, in case there is something (like a search) fixed up there.
@@ -368,9 +371,7 @@ const Dropdown = (props: DropdownProps): JSX.Element => {
                 className: classNames(child.props.className, {
                     'dropdown-item-selected': selected,
                     'dropdown-item-with-search-first-item': props.searchable === true && found === 0
-                }),
-                // TODO: this selected is wrong!
-                icon: selected && child.props.icon === undefined ? <DropdownCheckmark /> : child.props.icon
+                })
             })
             found += 1;
             return finalChild;
